fix: merge decodeStrictJSON unit tests for pkg-file-count cap
CI Build/Lint/API failed because handlers_common_test.go pushed pkg/api
to 16 files (limit 15). Consolidated helper unit tests into
server_test_helpers_test.go without changing coverage.
